Add success/failure log entries for install/update/uninstall events
This commit adds both "SUCCESSFUL" and "FAILED" messages to the gorilla.log file. Without the changes, Gorilla will log warnings if a "command" fails, but Gorilla does not write a log entry after running an install, update, or uninstall. Adding this change will make it easier to parse through the log and allow for quick analysis of just what Gorilla is installing, updating, or uninstalling along with the results of those events.