Prevent "update" items from installing if app is not already installed
Currently, Gorilla will install an item marked as a `managed_update` even if the app is not already installed. Meaning that, `managed_updates` and `managed_installs` are rendered equilvalents. This creates a problem for admins that wish to only manage the updates for a particular item. The changes in this diff correct the behavior. In addition to the fix(es), the commit also contains a couple edits to tailor debug log entries a bit, and make it look more standardized.