Remove windowStyle from PowerShell arguments
There is a potentially annoying problem with the way PowerShell scripts are executed via a `Script` check and running installers that are `ps1`. In the current state, and admin could have PowerShell open and run `gorilla` in a maximized or minimized window, and since there is a `windowStyle` argument of `Normal` the PowerShell window will always get resized to `Normal` unless the PowerShell window is already in that state/size.

This PR removes the `windowStyle` argument so that the window size/style does not change when `gorilla` is running. So, if an admin is running a maximized PowerShell window and runs `gorilla` the window will not resize.

The PR was tested via local run on Windows 10 after building.